Update readiness and startup probes
The readinessProb readinessThreshold and timeoutSeconds was too low, it was causing the pod to be restarted so incease the times allowed for timeout and a successful check.

Also the startpProbe of initialDelay of 600 seconds would mean that any change to the pod would take at least 10 minutes to take effect so reduce this to 60 seconds.